Lesson 8. Also fixed semi-colon issue.
Isomorphic! Had some difficulties because my dependencies were updated from the version LYR uses. Once I synced my node_modules to the LYR repo's. It fixed the problem for lesson 8 and also fixed the issue with the semi-colons in the styles.  Also had an issue with my binding of handleChange that was making it so I couldn't setState when the checkbox was checked. Fixed this.